Update: extract UI and update it to reuse existing scaffolding
