+++ WordPress/src/main/java/org/wordpress/android/ui/main/jetpack/staticposter/JetpackStaticPosterFragment.kt
@@ -4,15 +4,9 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.compose.foundation.layout.Box
-import androidx.compose.material.CircularProgressIndicator
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.platform.ComposeView
-import androidx.compose.ui.tooling.preview.Devices
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.lifecycleScope
@@ -20,9 +14,7 @@ import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
 import org.wordpress.android.ui.compose.theme.AppTheme
-import org.wordpress.android.ui.compose.utils.uiStringText
-import org.wordpress.android.ui.main.jetpack.staticposter.UiState.Content
-import org.wordpress.android.ui.main.jetpack.staticposter.UiState.Loading
+import org.wordpress.android.ui.main.jetpack.staticposter.compose.JetpackStaticPoster
 
 @AndroidEntryPoint
 class JetpackStaticPosterFragment : Fragment() {
@@ -32,11 +24,11 @@ class JetpackStaticPosterFragment : Fragment() {
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View = ComposeView(requireContext()).apply {
+    ) = ComposeView(requireContext()).apply {
         setContent {
             AppTheme {
                 val uiState by viewModel.uiState.collectAsState()
-                JetpackStaticPosterUi(uiState)
+                JetpackStaticPoster(uiState)
             }
         }
     }
@@ -64,37 +56,6 @@ class JetpackStaticPosterFragment : Fragment() {
             arguments = Bundle().apply {
                 putParcelable(ARG_PARCEL, parcel)
             }
-        }
-    }
-}
-
-@Composable
-private fun JetpackStaticPosterUi(uiState: UiState) {
-    Box {
-        when (uiState) {
-            is Content -> Content(uiState)
-            is Loading -> Loading()
-        }
-    }
-}
-
-@Composable
-private fun Content(uiState: Content) = with(uiState) {
-    Text(text = uiStringText(featureName))
-}
-
-@Composable
-private fun Loading() {
-    CircularProgressIndicator()
-}
-
-@Preview(showBackground = true, device = Devices.PIXEL_4_XL)
-@Composable
-private fun PreviewJetpackMoveScreen() {
-    AppTheme {
-        Box {
-            val uiState = UiData.STATS.toContentUiState()
-            JetpackStaticPosterUi(uiState)
         }
     }
 }
